bot/requirements.txt
fastapi
uvicorn
ccxt
pandas
pandas-ta
python-telegram-bot
requests
python-dotenv
